tasks/task_6/bot_v2_meta/requirements.txt
# Зависимости для Telegram-бота таймеров
# Обновлено для aiogram 3.21.0

# Основная библиотека для работы с Telegram Bot API
aiogram==3.21.0

# Библиотека для работы с .env файлами
python-dotenv==1.0.0
